Move `DecryptionError` to `common-crypto`
